Namecoin / AuxPoW: Remove calculate_merkle_root
It duplicates functionality of SPV.hash_merkle_root.
